refactor: improve extensibility
