+++ js/src/forum/extenders/extendIndexPage.tsx
@@ -1,8 +1,7 @@
 import app from 'flarum/forum/app';
 import { extend } from 'flarum/common/extend';
 import IndexPage from 'flarum/forum/components/IndexPage';
-import Dropdown from 'flarum/common/components/Dropdown';
-import Button from 'flarum/common/components/Button';
+import SolvedFilter from '../components/SolvedFilter';
 
 export default function extendIndexPage() {
   extend(IndexPage.prototype, 'sidebarItems', function (items) {
@@ -37,40 +36,8 @@ export default function extendIndexPage() {
       return;
     }
 
-    const options = ['all', 'solved', 'unsolved'];
-
     const selected = app.discussions.bestAnswer as unknown as number;
 
-    items.add(
-      'solved-filter',
-      Dropdown.component(
-        {
-          buttonClassName: 'Button',
-          label: app.translator.trans(
-            `fof-best-answer.forum.filter.${options[selected] || Object.keys(options).map((key) => options[Number(key)])[0]}_label`
-          ),
-          accessibleToggleLabel: app.translator.trans('fof-best-answer.forum.filter.accessible_label'),
-        },
-        Object.keys(options).map((value) => {
-          const label = options[Number(value)];
-          const active = (selected || Object.keys(options)[0]) === value;
-
-          return Button.component(
-            {
-              icon: active ? 'fas fa-check' : true,
-              active: active,
-              onclick: () => {
-                app.discussions.bestAnswer = value;
-                if (value === '0') {
-                  delete app.discussions.bestAnswer;
-                }
-                app.discussions.refresh();
-              },
-            },
-            app.translator.trans(`fof-best-answer.forum.filter.${label}_label`)
-          );
-        })
-      )
-    );
+    items.add('solved-filter', <SolvedFilter selected={selected} />);
   });
 }
